refactor(jsonl): split standard zstd physical stream
