+++ crates/ctx-history-jsonl/src/family/physical.rs
@@ -1,19 +1,32 @@
 use std::{
     fs::File,
-    io::{BufReader, Cursor, Read, Seek, SeekFrom, Write},
+    io::{BufReader, Cursor, Read, Seek, SeekFrom},
 };
 
 use sha2::{Digest, Sha256};
 
-use ctx_history_capture_runtime::{
-    SourceBackedRouteByteReservation, SourceBackedRouteResourceKind, SourceBackedRouteResources,
-};
+use ctx_history_capture_runtime::SourceBackedRouteResources;
 
 use super::{
     read_bounded_record, read_bounded_record_complete_and_prefix_sha256,
     read_bounded_record_complete_sha256, read_bounded_record_full_complete_and_prefix_sha256,
     read_bounded_record_unhashed, JsonlFamilyError, JsonlRecordFraming, JsonlResult,
     JsonlResumableSha256,
+};
+
+mod standard_zstd;
+#[cfg(test)]
+mod tests;
+
+#[cfg(any(test, feature = "test-support"))]
+pub use standard_zstd::set_after_standard_zstd_snapshot_hook;
+use standard_zstd::{
+    advance_standard_zstd_digest, decode_standard_zstd_jsonl, standard_zstd_physical_end,
+    StandardZstdBacking, StandardZstdLimits,
+};
+pub use standard_zstd::{
+    MAX_STANDARD_ZSTD_COMPRESSED_BYTES, MAX_STANDARD_ZSTD_DECOMPRESSED_BYTES,
+    MAX_STANDARD_ZSTD_PARALLEL_STREAMS, MAX_STANDARD_ZSTD_TEMP_BYTES_PER_LEAF,
 };
 
 const ZSTD_FRAME_MAGIC: [u8; 4] = [0x28, 0xb5, 0x2f, 0xfd];
@@ -23,19 +36,6 @@ const ZSTD_FRAME_MAGIC: [u8; 4] = [0x28, 0xb5, 0x2f, 0xfd];
 const MAX_ZSTD_DECODED_FRAME_BYTES: usize = 64 * 1024 * 1024;
 const MAX_ZSTD_FRAME_BYTES: usize = MAX_ZSTD_DECODED_FRAME_BYTES + (1024 * 1024);
 const MAX_ZSTD_WINDOW_LOG: u32 = 26;
-
-/// Hard physical bound for one provider-owned standard Zstandard JSONL stream.
-pub const MAX_STANDARD_ZSTD_COMPRESSED_BYTES: u64 = 256 * 1024 * 1024;
-/// Hard logical bound for one decoded standard Zstandard JSONL stream.
-pub const MAX_STANDARD_ZSTD_DECOMPRESSED_BYTES: u64 = 256 * 1024 * 1024;
-/// Compressed snapshot plus decoded spool retained by one standard stream.
-pub const MAX_STANDARD_ZSTD_TEMP_BYTES_PER_LEAF: u64 = 256 * 1024 * 1024;
-/// The production route scratch budget admits this many maximum-size leaves.
-pub const MAX_STANDARD_ZSTD_PARALLEL_STREAMS: usize = 4;
-const MAX_STANDARD_ZSTD_EXPANSION_RATIO: u64 = 256;
-const STANDARD_ZSTD_EXPANSION_SLACK_BYTES: u64 = 16 * 1024 * 1024;
-const MAX_STANDARD_ZSTD_WINDOW_LOG: u32 = 27;
-const STANDARD_ZSTD_COPY_BUFFER_BYTES: usize = 64 * 1024;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum JsonlPhysicalEncoding {
@@ -284,31 +284,6 @@ pub struct JsonlPhysicalStream<E: JsonlFamilyError> {
     record_buffer: Vec<u8>,
     incomplete_tail: bool,
     exhausted: bool,
-}
-
-#[derive(Debug)]
-struct StandardZstdBacking {
-    physical_reader: BufReader<File>,
-    logical_length: u64,
-    total_records: u64,
-    _scratch_reservations: Vec<SourceBackedRouteByteReservation>,
-}
-
-#[derive(Debug, Clone, Copy)]
-struct StandardZstdLimits {
-    compressed_bytes: u64,
-    decompressed_bytes: u64,
-    expansion_ratio: u64,
-    expansion_slack_bytes: u64,
-}
-
-impl StandardZstdLimits {
-    const PRODUCTION: Self = Self {
-        compressed_bytes: MAX_STANDARD_ZSTD_COMPRESSED_BYTES,
-        decompressed_bytes: MAX_STANDARD_ZSTD_DECOMPRESSED_BYTES,
-        expansion_ratio: MAX_STANDARD_ZSTD_EXPANSION_RATIO,
-        expansion_slack_bytes: STANDARD_ZSTD_EXPANSION_SLACK_BYTES,
-    };
 }
 
 impl<E: JsonlFamilyError> JsonlPhysicalStream<E> {
@@ -685,251 +660,6 @@ impl<E: JsonlFamilyError> JsonlPhysicalStream<E> {
     }
 }
 
-struct DecodedStandardZstd {
-    plaintext: File,
-    physical_reader: BufReader<File>,
-    logical_length: u64,
-    total_records: u64,
-    scratch_reservations: Vec<SourceBackedRouteByteReservation>,
-}
-
-fn decode_standard_zstd_jsonl<E: JsonlFamilyError>(
-    file: File,
-    compressed_length: u64,
-    framing: JsonlRecordFraming,
-    limits: StandardZstdLimits,
-    route_resources: Option<&SourceBackedRouteResources>,
-) -> JsonlResult<DecodedStandardZstd, E> {
-    if compressed_length == 0 {
-        return Err(invalid_standard_zstd::<E>("compressed stream is empty"));
-    }
-    if compressed_length > limits.compressed_bytes {
-        return Err(invalid_standard_zstd::<E>(
-            "compressed stream exceeds the bounded physical limit",
-        ));
-    }
-    if compressed_length > MAX_STANDARD_ZSTD_TEMP_BYTES_PER_LEAF {
-        return Err(invalid_standard_zstd::<E>(
-            "compressed snapshot exceeds the per-leaf temporary-volume limit",
-        ));
-    }
-    let ratio_bound = compressed_length
-        .saturating_mul(limits.expansion_ratio)
-        .saturating_add(limits.expansion_slack_bytes);
-    let decoded_bound = limits.decompressed_bytes.min(ratio_bound);
-    let compressed_reservation =
-        reserve_standard_zstd_scratch::<E>(route_resources, compressed_length)?;
-    let snapshot = snapshot_standard_zstd_prefix::<E>(file, compressed_length)?;
-    #[cfg(any(test, feature = "test-support"))]
-    run_after_standard_zstd_snapshot_hook();
-    let decoder_source = snapshot.try_clone()?;
-    let mut decoder = zstd::stream::read::Decoder::new(decoder_source).map_err(|error| {
-        E::invalid_payload(format!(
-            "invalid standard Zstandard JSONL stream header: {error}"
-        ))
-    })?;
-    decoder.window_log_max(MAX_STANDARD_ZSTD_WINDOW_LOG)?;
-    let mut plaintext = tempfile::tempfile()?;
-    let mut buffer = [0_u8; STANDARD_ZSTD_COPY_BUFFER_BYTES];
-    let mut logical_length = 0_u64;
-    let mut total_records = 0_u64;
-    let mut trailing_bytes = 0_u64;
-    let mut trailing_bytes_are_nul = true;
-    let mut scratch_reservations = Vec::new();
-    if let Some(reservation) = compressed_reservation {
-        scratch_reservations.push(reservation);
-    }
-    loop {
-        let read = decoder.read(&mut buffer).map_err(|error| {
-            E::invalid_payload(format!(
-                "corrupt or truncated Zstandard JSONL stream: {error}"
-            ))
-        })?;
-        if read == 0 {
-            break;
-        }
-        let read_u64 = u64::try_from(read)
-            .map_err(|_| E::system_invariant("decoded Zstandard chunk exceeds u64"))?;
-        logical_length = logical_length.checked_add(read_u64).ok_or_else(|| {
-            E::invalid_payload("decoded Zstandard JSONL length overflowed".to_owned())
-        })?;
-        if logical_length > decoded_bound {
-            return Err(invalid_standard_zstd::<E>(
-                "decoded stream exceeds the bounded decompression limit",
-            ));
-        }
-        if compressed_length.saturating_add(logical_length) > MAX_STANDARD_ZSTD_TEMP_BYTES_PER_LEAF
-        {
-            return Err(invalid_standard_zstd::<E>(
-                "compressed snapshot and decoded stream exceed the per-leaf temporary-volume limit",
-            ));
-        }
-        if let Some(reservation) = reserve_standard_zstd_scratch::<E>(route_resources, read_u64)? {
-            scratch_reservations.push(reservation);
-        }
-        for byte in &buffer[..read] {
-            if *byte == b'\n' {
-                total_records = total_records.checked_add(1).ok_or_else(|| {
-                    E::invalid_payload("decoded Zstandard JSONL record count overflowed".to_owned())
-                })?;
-                trailing_bytes = 0;
-                trailing_bytes_are_nul = true;
-            } else {
-                trailing_bytes = trailing_bytes.saturating_add(1);
-                trailing_bytes_are_nul &= *byte == 0;
-            }
-        }
-        plaintext.write_all(&buffer[..read])?;
-    }
-    drop(decoder.finish());
-    if logical_length == 0 {
-        return Err(invalid_standard_zstd::<E>(
-            "decoded stream contains no JSONL records",
-        ));
-    }
-    if trailing_bytes != 0 {
-        if framing.allows_terminal_nul_padding() && trailing_bytes_are_nul {
-            total_records = total_records.checked_add(1).ok_or_else(|| {
-                E::invalid_payload("decoded Zstandard JSONL record count overflowed".to_owned())
-            })?;
-        } else {
-            return Err(invalid_standard_zstd::<E>(
-                "decoded stream has a non-terminated JSONL tail",
-            ));
-        }
-    }
-    if total_records == 0 {
-        return Err(invalid_standard_zstd::<E>(
-            "decoded stream contains no complete JSONL records",
-        ));
-    }
-    plaintext.flush()?;
-    plaintext.seek(SeekFrom::Start(0))?;
-    let mut physical_reader = BufReader::new(snapshot);
-    physical_reader.seek(SeekFrom::Start(0))?;
-    Ok(DecodedStandardZstd {
-        plaintext,
-        physical_reader,
-        logical_length,
-        total_records,
-        scratch_reservations,
-    })
-}
-
-fn snapshot_standard_zstd_prefix<E: JsonlFamilyError>(
-    mut source: File,
-    compressed_length: u64,
-) -> JsonlResult<File, E> {
-    source.seek(SeekFrom::Start(0))?;
-    let mut snapshot = tempfile::tempfile()?;
-    let mut remaining = compressed_length;
-    let mut buffer = [0_u8; STANDARD_ZSTD_COPY_BUFFER_BYTES];
-    while remaining != 0 {
-        let take = usize::try_from(remaining.min(buffer.len() as u64))
-            .map_err(|_| E::system_invariant("Zstandard snapshot chunk exceeds usize"))?;
-        let read = source.read(&mut buffer[..take])?;
-        if read == 0 {
-            return Err(E::source_changed());
-        }
-        snapshot.write_all(&buffer[..read])?;
-        remaining = remaining.saturating_sub(read as u64);
-    }
-    snapshot.flush()?;
-    snapshot.seek(SeekFrom::Start(0))?;
-    Ok(snapshot)
-}
-
-fn reserve_standard_zstd_scratch<E: JsonlFamilyError>(
-    route_resources: Option<&SourceBackedRouteResources>,
-    bytes: u64,
-) -> JsonlResult<Option<SourceBackedRouteByteReservation>, E> {
-    let Some(route_resources) = route_resources else {
-        return Ok(None);
-    };
-    let bytes = usize::try_from(bytes).map_err(|_| {
-        E::invalid_payload("standard Zstandard scratch size exceeds usize".to_owned())
-    })?;
-    route_resources
-        .reserve(SourceBackedRouteResourceKind::LogicalSourceScratch, bytes)
-        .map(Some)
-        .map_err(|error| {
-            E::invalid_payload(format!(
-                "standard Zstandard JSONL temporary-volume budget unavailable: {error}"
-            ))
-        })
-}
-
-#[cfg(any(test, feature = "test-support"))]
-std::thread_local! {
-    static AFTER_STANDARD_ZSTD_SNAPSHOT_HOOK: std::cell::RefCell<Option<Box<dyn FnOnce()>>> =
-        const { std::cell::RefCell::new(None) };
-}
-
-#[cfg(any(test, feature = "test-support"))]
-pub fn set_after_standard_zstd_snapshot_hook(hook: impl FnOnce() + 'static) {
-    AFTER_STANDARD_ZSTD_SNAPSHOT_HOOK.with(|slot| {
-        assert!(
-            slot.borrow().is_none(),
-            "standard Zstandard snapshot hook is already installed"
-        );
-        *slot.borrow_mut() = Some(Box::new(hook));
-    });
-}
-
-#[cfg(any(test, feature = "test-support"))]
-fn run_after_standard_zstd_snapshot_hook() {
-    let hook = AFTER_STANDARD_ZSTD_SNAPSHOT_HOOK.with(|slot| slot.borrow_mut().take());
-    if let Some(hook) = hook {
-        hook();
-    }
-}
-
-fn standard_zstd_physical_end(
-    current: u64,
-    physical_length: u64,
-    logical_end: u64,
-    logical_length: u64,
-    next_ordinal: u64,
-    total_records: u64,
-) -> u64 {
-    if logical_end >= logical_length || next_ordinal >= total_records {
-        return physical_length;
-    }
-    let proportional =
-        (u128::from(logical_end) * u128::from(physical_length) / u128::from(logical_length)) as u64;
-    current
-        .max(proportional)
-        .min(physical_length.saturating_sub(1))
-}
-
-fn advance_standard_zstd_digest<E: JsonlFamilyError>(
-    reader: &mut BufReader<File>,
-    digest: &mut JsonlPhysicalDigest,
-    start: u64,
-    end: u64,
-    complete: bool,
-) -> JsonlResult<(), E> {
-    if end < start {
-        return Err(E::system_invariant(
-            "standard Zstandard JSONL physical progress regressed",
-        ));
-    }
-    let mut remaining = end - start;
-    let mut buffer = [0_u8; STANDARD_ZSTD_COPY_BUFFER_BYTES];
-    while remaining != 0 {
-        let take = usize::try_from(remaining.min(buffer.len() as u64))
-            .map_err(|_| E::system_invariant("Zstandard digest chunk exceeds usize"))?;
-        reader.read_exact(&mut buffer[..take])?;
-        digest.update_physical_unit(&buffer[..take], complete);
-        remaining = remaining.saturating_sub(take as u64);
-    }
-    Ok(())
-}
-
-fn invalid_standard_zstd<E: JsonlFamilyError>(detail: &str) -> E {
-    E::invalid_payload(format!("invalid standard Zstandard JSONL: {detail}"))
-}
-
 struct JsonlDecodedPhysicalUnit {
     complete: bool,
     terminal_nul_padding: bool,
@@ -1089,422 +819,4 @@ fn read_frame_part<E: JsonlFamilyError>(
 
 fn invalid_zstd<E: JsonlFamilyError>(detail: &str) -> E {
     E::invalid_payload(format!("invalid checksummed Zstandard JSONL: {detail}"))
-}
-
-#[cfg(test)]
-mod tests {
-    use ctx_history_source_io::SourceIoError;
-    use sha2::{Digest, Sha256};
-
-    use super::*;
-
-    #[test]
-    fn stream_tracks_complete_prefix_tail_and_rollback() {
-        let temp = tempfile::tempdir().unwrap();
-        let path = temp.path().join("records.jsonl");
-        let contents = b"one\ntwo\nincomplete";
-        std::fs::write(&path, contents).unwrap();
-        let mut stream = JsonlPhysicalStream::open(
-            File::open(&path).unwrap(),
-            contents.len() as u64,
-            0,
-            0,
-            JsonlRecordFraming::ordinary(),
-            JsonlPhysicalDigest::full_and_complete(
-                JsonlResumableSha256::new(),
-                JsonlResumableSha256::new(),
-            ),
-            || SourceIoError::SourceChangedDuringCapture,
-        )
-        .unwrap();
-
-        let first = stream.next_record().unwrap().unwrap();
-        assert_eq!(stream.record_bytes(first), b"one");
-        let after_first = stream.position();
-        let second = stream.next_record().unwrap().unwrap();
-        assert_eq!(stream.record_bytes(second), b"two");
-        stream.restore(after_first).unwrap();
-        assert_eq!(stream.next_record().unwrap().unwrap(), second);
-        let tail = stream.next_record().unwrap().unwrap();
-        assert!(!tail.complete);
-        assert_eq!(stream.record_bytes(tail), b"incomplete");
-        let after_tail = stream.position();
-        assert!(stream.next_record().unwrap().is_none());
-        stream.restore(after_tail).unwrap();
-        assert!(stream.next_record().unwrap().is_none());
-        assert_eq!(stream.complete_prefix_end(), 8);
-        assert_eq!(stream.next_physical_ordinal(), 2);
-        assert!(!stream.terminal());
-        let digest = stream.digest();
-        let complete = digest.complete_hasher().digest();
-        let expected_complete: [u8; 32] = Sha256::digest(&contents[..8]).into();
-        assert_eq!(complete, expected_complete);
-        let full = digest.full_hasher().unwrap().digest();
-        let expected_full: [u8; 32] = Sha256::digest(contents).into();
-        assert_eq!(full, expected_full);
-    }
-
-    fn checksummed_frame(plaintext: &[u8]) -> Vec<u8> {
-        let mut encoder = zstd::stream::Encoder::new(Vec::new(), 1).unwrap();
-        encoder.include_checksum(true).unwrap();
-        std::io::Write::write_all(&mut encoder, plaintext).unwrap();
-        encoder.finish().unwrap()
-    }
-
-    fn standard_stream(plaintext: &[u8]) -> Vec<u8> {
-        zstd::stream::encode_all(Cursor::new(plaintext), 1).unwrap()
-    }
-
-    #[test]
-    fn standard_zstd_jsonl_streams_bounded_records_and_rolls_back() {
-        let temp = tempfile::tempdir().unwrap();
-        let path = temp.path().join("rollout.jsonl.zst");
-        let plaintext = b"{\"type\":\"session_meta\"}\n{\"type\":\"response_item\"}\n";
-        let compressed = standard_stream(plaintext);
-        std::fs::write(&path, &compressed).unwrap();
-        let mut stream = JsonlPhysicalStream::open_with_encoding(
-            File::open(&path).unwrap(),
-            compressed.len() as u64,
-            0,
-            0,
-            JsonlPhysicalEncoding::StandardZstdJsonl,
-            JsonlRecordFraming::ordinary(),
-            JsonlPhysicalDigest::full_and_complete(
-                JsonlResumableSha256::new(),
-                JsonlResumableSha256::new(),
-            ),
-            || SourceIoError::SourceChangedDuringCapture,
-        )
-        .unwrap();
-
-        let first = stream.next_record().unwrap().unwrap();
-        assert_eq!(stream.record_bytes(first), b"{\"type\":\"session_meta\"}");
-        let after_first = stream.position();
-        let second = stream.next_record().unwrap().unwrap();
-        assert_eq!(stream.record_bytes(second), b"{\"type\":\"response_item\"}");
-        stream.restore(after_first).unwrap();
-        assert_eq!(stream.next_record().unwrap().unwrap(), second);
-        assert!(stream.next_record().unwrap().is_none());
-        assert!(stream.terminal());
-        assert_eq!(stream.complete_prefix_end(), compressed.len() as u64);
-        assert_eq!(stream.next_physical_ordinal(), 2);
-        assert_eq!(
-            stream.digest().complete_hasher().digest(),
-            <[u8; 32]>::from(Sha256::digest(&compressed))
-        );
-        assert_eq!(
-            stream.digest().full_hasher().unwrap().digest(),
-            <[u8; 32]>::from(Sha256::digest(&compressed))
-        );
-    }
-
-    #[test]
-    fn standard_zstd_jsonl_rejects_corrupt_truncated_and_unterminated_streams() {
-        let temp = tempfile::tempdir().unwrap();
-        let valid = standard_stream(b"{}\n");
-        let cases = [
-            ("corrupt", {
-                let mut value = valid.clone();
-                value[0] ^= 0xff;
-                value
-            }),
-            ("truncated", valid[..valid.len().saturating_sub(2)].to_vec()),
-            ("unterminated", standard_stream(b"{}")),
-        ];
-        for (label, bytes) in cases {
-            let path = temp.path().join(format!("{label}.jsonl.zst"));
-            std::fs::write(&path, &bytes).unwrap();
-            let result = JsonlPhysicalStream::open_with_encoding(
-                File::open(&path).unwrap(),
-                bytes.len() as u64,
-                0,
-                0,
-                JsonlPhysicalEncoding::StandardZstdJsonl,
-                JsonlRecordFraming::ordinary(),
-                JsonlPhysicalDigest::complete(JsonlResumableSha256::new()),
-                || SourceIoError::SourceChangedDuringCapture,
-            );
-            assert!(result.is_err(), "{label} stream was accepted");
-        }
-    }
-
-    #[test]
-    fn standard_zstd_jsonl_enforces_physical_logical_and_expansion_limits() {
-        let temp = tempfile::tempdir().unwrap();
-        let plaintext = b"{}\n".repeat(4096);
-        let compressed = standard_stream(&plaintext);
-        let path = temp.path().join("bounded.jsonl.zst");
-        std::fs::write(&path, &compressed).unwrap();
-        let base = StandardZstdLimits {
-            compressed_bytes: compressed.len() as u64,
-            decompressed_bytes: plaintext.len() as u64,
-            expansion_ratio: u64::MAX,
-            expansion_slack_bytes: u64::MAX,
-        };
-
-        for limits in [
-            StandardZstdLimits {
-                compressed_bytes: (compressed.len() as u64).saturating_sub(1),
-                ..base
-            },
-            StandardZstdLimits {
-                decompressed_bytes: (plaintext.len() as u64).saturating_sub(1),
-                ..base
-            },
-            StandardZstdLimits {
-                expansion_ratio: 1,
-                expansion_slack_bytes: 0,
-                ..base
-            },
-        ] {
-            let result = decode_standard_zstd_jsonl::<SourceIoError>(
-                File::open(&path).unwrap(),
-                compressed.len() as u64,
-                JsonlRecordFraming::ordinary(),
-                limits,
-                None,
-            );
-            assert!(result.is_err());
-        }
-    }
-
-    #[test]
-    fn standard_zstd_snapshot_hash_and_decode_use_only_the_certified_prefix() {
-        use std::fs::OpenOptions;
-
-        for overwrite in [false, true] {
-            let temp = tempfile::tempdir().unwrap();
-            let path = temp.path().join("bounded-prefix.jsonl.zst");
-            let plaintext = b"{\"type\":\"session_meta\"}\n{\"type\":\"response_item\"}\n";
-            let compressed = standard_stream(plaintext);
-            std::fs::write(&path, &compressed).unwrap();
-            let hook_path = path.clone();
-            let appended = standard_stream(b"{\"type\":\"appended\"}\n");
-            let hook_compressed = compressed.clone();
-            let hook_appended = appended.clone();
-            set_after_standard_zstd_snapshot_hook(move || {
-                if overwrite {
-                    let mut replacement = hook_compressed;
-                    replacement[0] ^= 0xff;
-                    OpenOptions::new()
-                        .write(true)
-                        .open(&hook_path)
-                        .unwrap()
-                        .write_all(&replacement)
-                        .unwrap();
-                } else {
-                    OpenOptions::new()
-                        .append(true)
-                        .open(&hook_path)
-                        .unwrap()
-                        .write_all(&hook_appended)
-                        .unwrap();
-                }
-            });
-
-            let mut stream = JsonlPhysicalStream::open_with_encoding(
-                File::open(&path).unwrap(),
-                compressed.len() as u64,
-                0,
-                0,
-                JsonlPhysicalEncoding::StandardZstdJsonl,
-                JsonlRecordFraming::ordinary(),
-                JsonlPhysicalDigest::complete(JsonlResumableSha256::new()),
-                || SourceIoError::SourceChangedDuringCapture,
-            )
-            .unwrap();
-            let mut records = Vec::new();
-            while let Some(record) = stream.next_record().unwrap() {
-                records.push(stream.record_bytes(record).to_vec());
-            }
-            assert_eq!(
-                records,
-                [
-                    b"{\"type\":\"session_meta\"}".to_vec(),
-                    b"{\"type\":\"response_item\"}".to_vec()
-                ]
-            );
-            assert_eq!(
-                stream.digest().complete_hasher().digest(),
-                <[u8; 32]>::from(Sha256::digest(&compressed))
-            );
-            let current = std::fs::read(&path).unwrap();
-            if overwrite {
-                assert_ne!(&current[..compressed.len()], compressed.as_slice());
-            } else {
-                assert_eq!(&current[..compressed.len()], compressed.as_slice());
-                assert_eq!(&current[compressed.len()..], appended.as_slice());
-            }
-        }
-    }
-
-    #[test]
-    fn standard_zstd_parallel_scratch_budget_fails_closed_and_releases() {
-        let temp = tempfile::tempdir().unwrap();
-        let path = temp.path().join("parallel-budget.jsonl.zst");
-        let plaintext = b"{}\n".repeat(4096);
-        let compressed = standard_stream(&plaintext);
-        std::fs::write(&path, &compressed).unwrap();
-        let one_stream_bytes = compressed.len().saturating_add(plaintext.len());
-        let resources = SourceBackedRouteResources::for_test(2, 0, one_stream_bytes as u64);
-        let open = || {
-            JsonlPhysicalStream::open_with_encoding_and_resources(
-                File::open(&path).unwrap(),
-                compressed.len() as u64,
-                0,
-                0,
-                JsonlPhysicalEncoding::StandardZstdJsonl,
-                JsonlRecordFraming::ordinary(),
-                JsonlPhysicalDigest::complete(JsonlResumableSha256::new()),
-                || SourceIoError::SourceChangedDuringCapture,
-                Some(&resources),
-            )
-        };
-
-        let first = open().unwrap();
-        assert_eq!(
-            resources.live_bytes(SourceBackedRouteResourceKind::LogicalSourceScratch),
-            one_stream_bytes as u64
-        );
-        assert!(
-            open().is_err(),
-            "a concurrent second spool exceeded the shared budget"
-        );
-        drop(first);
-        assert_eq!(
-            resources.live_bytes(SourceBackedRouteResourceKind::LogicalSourceScratch),
-            0
-        );
-        assert!(open().is_ok());
-    }
-
-    #[test]
-    fn zstd_stream_decodes_concatenated_checksummed_frames_and_rolls_back() {
-        let temp = tempfile::tempdir().unwrap();
-        let path = temp.path().join("session.jsonl.zstd");
-        let first_bytes = checksummed_frame(b"{\"type\":\"session\"}\n");
-        let second_bytes =
-            checksummed_frame(b"{\"type\":\"user/message\"}\n{\"type\":\"assistant/message\"}\n");
-        let contents = [first_bytes.as_slice(), second_bytes.as_slice()].concat();
-        std::fs::write(&path, &contents).unwrap();
-        let mut stream = JsonlPhysicalStream::open_with_encoding(
-            File::open(&path).unwrap(),
-            contents.len() as u64,
-            0,
-            0,
-            JsonlPhysicalEncoding::ChecksummedZstdFrames,
-            JsonlRecordFraming::ordinary(),
-            JsonlPhysicalDigest::full_and_complete(
-                JsonlResumableSha256::new(),
-                JsonlResumableSha256::new(),
-            ),
-            || SourceIoError::SourceChangedDuringCapture,
-        )
-        .unwrap();
-
-        let header = stream.next_record().unwrap().unwrap();
-        assert_eq!(stream.record_bytes(header), b"{\"type\":\"session\"}\n");
-        let after_header = stream.position();
-        let batch = stream.next_record().unwrap().unwrap();
-        assert_eq!(
-            stream.record_bytes(batch),
-            b"{\"type\":\"user/message\"}\n{\"type\":\"assistant/message\"}\n"
-        );
-        stream.restore(after_header).unwrap();
-        assert_eq!(stream.next_record().unwrap().unwrap(), batch);
-        assert!(stream.next_record().unwrap().is_none());
-        assert!(stream.terminal());
-        assert_eq!(stream.complete_prefix_end(), contents.len() as u64);
-        assert_eq!(stream.next_physical_ordinal(), 2);
-        assert_eq!(
-            stream.digest().complete_hasher().digest(),
-            <[u8; 32]>::from(Sha256::digest(&contents))
-        );
-    }
-
-    #[test]
-    fn zstd_stream_omits_torn_frame_and_rejects_corrupt_or_unchecked_frames() {
-        let temp = tempfile::tempdir().unwrap();
-        let complete = checksummed_frame(b"{\"type\":\"session\"}\n");
-        let mut torn = checksummed_frame(b"{\"type\":\"user/message\"}\n");
-        torn.truncate(torn.len().saturating_sub(2));
-        let contents = [complete.as_slice(), torn.as_slice()].concat();
-        let path = temp.path().join("torn.jsonl.zstd");
-        std::fs::write(&path, &contents).unwrap();
-        let mut stream = JsonlPhysicalStream::open_with_encoding(
-            File::open(&path).unwrap(),
-            contents.len() as u64,
-            0,
-            0,
-            JsonlPhysicalEncoding::ChecksummedZstdFrames,
-            JsonlRecordFraming::ordinary(),
-            JsonlPhysicalDigest::full_and_complete(
-                JsonlResumableSha256::new(),
-                JsonlResumableSha256::new(),
-            ),
-            || SourceIoError::SourceChangedDuringCapture,
-        )
-        .unwrap();
-        assert!(stream.next_record().unwrap().unwrap().complete);
-        assert!(!stream.next_record().unwrap().unwrap().complete);
-        assert!(!stream.terminal());
-        assert_eq!(stream.complete_prefix_end(), complete.len() as u64);
-        assert_eq!(
-            stream.digest().full_hasher().unwrap().digest(),
-            <[u8; 32]>::from(Sha256::digest(&contents))
-        );
-
-        let mut corrupt = complete.clone();
-        let last = corrupt.len().saturating_sub(1);
-        corrupt[last] ^= 0xff;
-        let corrupt_path = temp.path().join("corrupt.jsonl.zstd");
-        std::fs::write(&corrupt_path, &corrupt).unwrap();
-        let mut corrupt_stream = JsonlPhysicalStream::open_with_encoding(
-            File::open(&corrupt_path).unwrap(),
-            corrupt.len() as u64,
-            0,
-            0,
-            JsonlPhysicalEncoding::ChecksummedZstdFrames,
-            JsonlRecordFraming::ordinary(),
-            JsonlPhysicalDigest::complete(JsonlResumableSha256::new()),
-            || SourceIoError::SourceChangedDuringCapture,
-        )
-        .unwrap();
-        assert!(corrupt_stream.next_record().is_err());
-
-        let mut encoder = zstd::stream::Encoder::new(Vec::new(), 1).unwrap();
-        std::io::Write::write_all(&mut encoder, b"{}\n").unwrap();
-        let unchecked = encoder.finish().unwrap();
-        let unchecked_path = temp.path().join("unchecked.jsonl.zstd");
-        std::fs::write(&unchecked_path, &unchecked).unwrap();
-        let mut unchecked_stream = JsonlPhysicalStream::open_with_encoding(
-            File::open(&unchecked_path).unwrap(),
-            unchecked.len() as u64,
-            0,
-            0,
-            JsonlPhysicalEncoding::ChecksummedZstdFrames,
-            JsonlRecordFraming::ordinary(),
-            JsonlPhysicalDigest::complete(JsonlResumableSha256::new()),
-            || SourceIoError::SourceChangedDuringCapture,
-        )
-        .unwrap();
-        assert!(unchecked_stream.next_record().is_err());
-
-        let mut reserved = complete.clone();
-        reserved[4] |= 0x10;
-        let reserved_path = temp.path().join("reserved.jsonl.zstd");
-        std::fs::write(&reserved_path, &reserved).unwrap();
-        let mut reserved_stream = JsonlPhysicalStream::open_with_encoding(
-            File::open(&reserved_path).unwrap(),
-            reserved.len() as u64,
-            0,
-            0,
-            JsonlPhysicalEncoding::ChecksummedZstdFrames,
-            JsonlRecordFraming::ordinary(),
-            JsonlPhysicalDigest::complete(JsonlResumableSha256::new()),
-            || SourceIoError::SourceChangedDuringCapture,
-        )
-        .unwrap();
-        assert!(reserved_stream.next_record().is_err());
-    }
 }
